Fixed liveMap and plot order in MonolithicJs Fixed dispatching mouseEvents in DomCanvasControl Fixed LiveMap position on plot in AWT, when using title and facet grid

+++ plot-builder/src/commonMain/kotlin/jetbrains/datalore/plot/builder/layout/PlotLayoutBase.kt
@@ -6,13 +6,13 @@
 package jetbrains.datalore.plot.builder.layout
 
 internal abstract class PlotLayoutBase : PlotLayout {
-    protected var paddingTop: Double = 0.toDouble()
+    protected var paddingTop: Double = 0.0
         private set
-    protected var paddingRight: Double = 0.toDouble()
+    protected var paddingRight: Double = 0.0
         private set
-    protected var paddingBottom: Double = 0.toDouble()
+    protected var paddingBottom: Double = 0.0
         private set
-    protected var paddingLeft: Double = 0.toDouble()
+    protected var paddingLeft: Double = 0.0
         private set
 
     override fun setPadding(top: Double, right: Double, bottom: Double, left: Double) {
